Models and Repositories refactor
