+++ api/src/main/java/com/example/demo/ApiApplication.java
@@ -8,7 +8,6 @@ public class ApiApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(ApiApplication.class, args);
-		System.out.println("Hello World");
 	}
 
 }
