--- api/src/main/java/com/example/demo/Models/UserAddressID.java
@@ -1,76 +1,0 @@
-package com.example.demo.Models;
-
-import java.io.Serializable;
-import java.util.Set;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import javax.persistence.OneToMany;
-
-@Embeddable
-public class UserAddressID implements Serializable {
-
-    @Column(name = "user_id")
-    private Long userId;
-
-    @Column(name = "address_id")
-    private Long addressId;
-
-    public UserAddressID() {
-    }
-
-    public UserAddressID(Long userId, Long addressId) {
-        this.userId = userId;
-        this.addressId = addressId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public Long getAddressId() {
-        return addressId;
-    }
-
-    public void setAddressId(Long addressId) {
-        this.addressId = addressId;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((addressId == null) ? 0 : addressId.hashCode());
-        result = prime * result + ((userId == null) ? 0 : userId.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        UserAddressID other = (UserAddressID) obj;
-        if (addressId == null) {
-            if (other.addressId != null)
-                return false;
-        } else if (!addressId.equals(other.addressId))
-            return false;
-        if (userId == null) {
-            if (other.userId != null)
-                return false;
-        } else if (!userId.equals(other.userId))
-            return false;
-        return true;
-    }
-    
-
-    
-}
