+++ api/src/main/java/com/example/demo/models/Address.java
@@ -1,0 +1,96 @@
+package com.example.demo.models;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
+
+import javax.persistence.Table;
+
+@Table(name="address")
+@Entity
+public class Address {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column
+    @NotNull(message = "pais é obrigatório")
+    private String country;
+
+    @Column
+    @NotNull(message = "zipcode é obrigatório")
+    private String zipcode;
+
+    @Column
+    @NotNull(message = "city é obrigatório")
+    private String city;
+
+    @Column
+    @NotNull(message = "address é obrigatório")
+    private String address;
+
+    public Address() {       
+    }
+
+    public Address(String country, String zipcode, String city, String address) {
+        this.country = country;
+        this.zipcode = zipcode;
+        this.city = city;
+        this.address = address;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getCountry() {
+        return country;
+    }
+
+    public void setCountry(String country) {
+        this.country = country;
+    }
+
+    public String getZipcode() {
+        return zipcode;
+    }
+
+    public void setZipcode(String zipcode) {
+        this.zipcode = zipcode;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    @Override
+    public String toString() {
+        return "{" +
+            " id='" + getId() + "'" +
+            ", country='" + getCountry() + "'" +
+            ", zipcode='" + getZipcode() + "'" +
+            ", city='" + getCity() + "'" +
+            ", address='" + getAddress() + "'" +
+            "}";
+    }    
+    
+}
