+++ api/src/main/java/com/example/demo/models/CartList.java
@@ -1,4 +1,4 @@
-package com.example.demo.Models;
+package com.example.demo.models;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -8,13 +8,9 @@ import javax.persistence.MapsId;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
-import org.springframework.beans.factory.annotation.Value;
-
-import lombok.Builder.Default;
-
-@Table(name="savedList")
+@Table(name="cartList")
 @Entity
-public class SavedList {
+public class CartList {
     @Id
     @Column(name = "productList_id")
     private Long id;
@@ -25,52 +21,24 @@ public class SavedList {
     @JoinColumn(name = "productList_id")
     private ProductList productList;
 
-    @Column
-    @Value("")
-    private String name;
 
-    public SavedList() {
+    public CartList() {
     }
 
 
-    
-
-
-    public SavedList(ProductList productList, String name) {
+    public CartList(ProductList productList) {
         this.productList = productList;
-        this.name = name;
     }
-
-
-
 
 
     public ProductList getProductList() {
         return productList;
-    }
-    
-
-    public Long getId() {
-        return id;
-    }
-
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-
-    public String getName() {
-        return name;
-    }
-
-
-    public void setName(String name) {
-        this.name = name;
     }
 
 
     public void setProductList(ProductList productList) {
         this.productList = productList;
     }
+
+    
 }
