+++ api/src/main/java/com/example/demo/models/OrderList.java
@@ -1,4 +1,5 @@
-package com.example.demo.Models;
+package com.example.demo.models;
+
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -6,15 +7,12 @@ import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.MapsId;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
-import javax.persistence.PrimaryKeyJoinColumn;
 import javax.validation.constraints.NotNull;
 
 import org.hibernate.annotations.CreationTimestamp;
@@ -37,7 +35,7 @@ public class OrderList {
     private ProductList productList;
 
     @OneToMany(mappedBy="orderList")
-    private Set<OrderProductItem> orderProductItem;
+    private Set<OrderProductItem> orderProductItems;
 
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "address_id", referencedColumnName = "id" , nullable = false)
@@ -46,8 +44,6 @@ public class OrderList {
     @ManyToOne
     @JoinColumn(name="store_id", nullable=false)
     private Store store;
-
-    
 
     @Column(name = "delivery_id")
     @NotNull(message = "deliveryId é obrigatório")
@@ -63,42 +59,40 @@ public class OrderList {
     private Date orderTimestamp;
 
     public OrderList() {
-        this.orderProductItem= new HashSet<>();
+        this.orderProductItems = new HashSet<>();
     }
 
-    public OrderList(ProductList productList, Set<OrderProductItem> productListItems) {
+    public OrderList(ProductList productList, Address address, Store store, Long deliveryId, Long deliveryTimestamp) {
         this.productList = productList;
-        this.orderProductItem = productListItems;
-        this.orderProductItem= new HashSet<>();
-    }
-
-    
-
-    public OrderList(ProductList productList, Set<OrderProductItem> orderProductItem, Address address, Store store,
-             Long deliveryId,
-            Long deliveryTimestamp) {
-        this.productList = productList;
-        this.orderProductItem = orderProductItem;
         this.address = address;
         this.store = store;
         this.deliveryId = deliveryId;
         this.deliveryTimestamp = deliveryTimestamp;
+        this.orderProductItems = new HashSet<>();
     }
 
     public ProductList getProductList() {
         return productList;
     }
     
-    public Set<OrderProductItem> getOrderProductItem() {
-        return orderProductItem;
+    public Long getId() {
+        return this.id;
     }
 
-    public void setOrderProductItem(Set<OrderProductItem> orderProductItem) {
-        this.orderProductItem = orderProductItem;
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setProductList(ProductList productList) {
+        this.productList = productList;
+    }
+
+    public Set<OrderProductItem> getOrderProductItems() {
+        return this.orderProductItems;
     }
 
     public Address getAddress() {
-        return address;
+        return this.address;
     }
 
     public void setAddress(Address address) {
@@ -106,15 +100,15 @@ public class OrderList {
     }
 
     public Store getStore() {
-        return store;
+        return this.store;
     }
 
     public void setStore(Store store) {
         this.store = store;
     }
-    
+
     public Long getDeliveryId() {
-        return deliveryId;
+        return this.deliveryId;
     }
 
     public void setDeliveryId(Long deliveryId) {
@@ -122,7 +116,7 @@ public class OrderList {
     }
 
     public Long getDeliveryTimestamp() {
-        return deliveryTimestamp;
+        return this.deliveryTimestamp;
     }
 
     public void setDeliveryTimestamp(Long deliveryTimestamp) {
@@ -130,32 +124,11 @@ public class OrderList {
     }
 
     public Date getOrderTimestamp() {
-        return orderTimestamp;
+        return this.orderTimestamp;
     }
 
     public void setOrderTimestamp(Date orderTimestamp) {
         this.orderTimestamp = orderTimestamp;
     }
-
-    public void setProductList(ProductList productList) {
-        this.productList = productList;
-    }
-
-    public Set<OrderProductItem> getProductListItems() {
-        return orderProductItem;
-    }
-
-    public void setProductListItems(Set<OrderProductItem> productListItems) {
-        this.orderProductItem = productListItems;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-    
     
 }
