+++ api/src/main/java/com/example/demo/models/OrderProductItem.java
@@ -1,9 +1,8 @@
-package com.example.demo.Models;
+package com.example.demo.models;
 
 import javax.persistence.Column;
 import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.MapsId;
