+++ api/src/main/java/com/example/demo/models/Product.java
@@ -1,4 +1,4 @@
-package com.example.demo.Models;
+package com.example.demo.models;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
