+++ api/src/main/java/com/example/demo/repository/AddressRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.Repository;
+package com.example.demo.repository;
 
-import com.example.demo.Models.Address;
+import com.example.demo.models.Address;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
