+++ api/src/main/java/com/example/demo/repository/CartListRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.Repository;
+package com.example.demo.repository;
 
-import com.example.demo.Models.CartList;
+import com.example.demo.models.CartList;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
