+++ api/src/main/java/com/example/demo/repository/OrderListRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.Repository;
+package com.example.demo.repository;
 
-import com.example.demo.Models.OrderList;
+import com.example.demo.models.OrderList;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
