+++ api/src/main/java/com/example/demo/repository/ProductListItemRepository.java
@@ -1,7 +1,7 @@
-package com.example.demo.Repository;
+package com.example.demo.repository;
 
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.ProductListItemId;
+import com.example.demo.models.ProductListItem;
+import com.example.demo.models.ProductListItemId;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
