+++ api/src/main/java/com/example/demo/repository/ProductListRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo.Repository;
+package com.example.demo.repository;
 
-import com.example.demo.Models.ProductList;
+import com.example.demo.models.ProductList;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
