+++ api/src/main/java/com/example/demo/repository/UserAddressRepository.java
@@ -1,0 +1,12 @@
+package com.example.demo.repository;
+
+import com.example.demo.models.UserAddress;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface UserAddressRepository extends JpaRepository<UserAddress, Long> {
+  
+}
