+++ api/src/test/java/com/example/demo/model/AddressTest.java
@@ -1,11 +1,7 @@
 package com.example.demo.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.Address;
-import com.example.demo.Models.UserAddress;
+import com.example.demo.models.Address;
 
 import org.junit.jupiter.api.Test;
 public class AddressTest {
@@ -17,34 +13,23 @@ public class AddressTest {
         x.setAddress("Rua das Pombas");
         x.setCity("Aveiro");
         x.setCountry("Portugal");
-        x.setOrderList(null);
         x.setId(1);
-        x.setStore(null);
         x.setZipcode("1204-322");
-        UserAddress u_adress= new UserAddress();
-        x.setUserAddress(u_adress);
 
         assertEquals("Rua das Pombas", x.getAddress());
         assertEquals("Aveiro", x.getCity());
         assertEquals("Portugal", x.getCountry());
         assertEquals("1204-322", x.getZipcode());
-        assertEquals(null, x.getOrderList());
         assertEquals(1, x.getId());
-        assertEquals(null, x.getStore());
-
     }
+
     @Test
     void testWhenCreateValidAddressWithConstuctorWithFullParametersThenReturnAddress(){
-        UserAddress u_adress= new UserAddress();
-        Address x = new Address("Portugal", "1204-322", "Aveiro", "Rua das Pombas", null, null, u_adress);
-
+        Address x = new Address("Portugal", "1204-322", "Aveiro", "Rua das Pombas");
 
         assertEquals("Rua das Pombas", x.getAddress());
         assertEquals("Aveiro", x.getCity());
         assertEquals("Portugal", x.getCountry());
         assertEquals("1204-322", x.getZipcode());
-        assertEquals(null, x.getOrderList());
-        assertEquals(null, x.getStore());
-
     }
 }
