+++ api/src/test/java/com/example/demo/model/CategoryTest.java
@@ -1,15 +1,8 @@
 package com.example.demo.model;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.UserAddress;
+import com.example.demo.models.Category;
 
 import org.junit.jupiter.api.Test;
 public class CategoryTest {
@@ -19,22 +12,17 @@ public class CategoryTest {
         cat.setActive(false);
         cat.setId(1);
         cat.setName("Vegetais");
-        Set<Product> x = new HashSet();
-        cat.setProduct(x);
 
         assertEquals(1, cat.getId());
         assertEquals("Vegetais", cat.getName());
         assertEquals(false, cat.getActive());
-        assertEquals(0, cat.getProduct().size());
     }
     @Test
     void testWhenCreateValidCategoryWithConstuctorThenReturnCategory(){
-        Set<Product> x = new HashSet();
-        Category cat = new Category("Vegetais", false, x);
+        Category cat = new Category("Vegetais", false);
 
         assertEquals("Vegetais", cat.getName());
         assertEquals(false, cat.getActive());
-        assertEquals(0, cat.getProduct().size());
     }
 
     @Test
@@ -43,7 +31,6 @@ public class CategoryTest {
 
         assertEquals("Vegetais", cat.getName());
         assertEquals(false, cat.getActive());
-        assertEquals(0, cat.getProduct().size());
     }
 
 }
