+++ api/src/test/java/com/example/demo/model/OrderProductItemTest.java
@@ -1,20 +1,10 @@
 package com.example.demo.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.sql.Date;
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.OrderProductItemId;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.Store;
-import com.example.demo.Models.UserAddress;
+import com.example.demo.models.OrderList;
+import com.example.demo.models.OrderProductItem;
+import com.example.demo.models.OrderProductItemId;
+import com.example.demo.models.Product;
 
 import org.junit.jupiter.api.Test;
 public class OrderProductItemTest {
