+++ api/src/test/java/com/example/demo/model/ProductListItemTest.java
@@ -1,19 +1,10 @@
 package com.example.demo.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.ProductListItemId;
-import com.example.demo.Models.SavedList;
-import com.example.demo.Models.User;
+import com.example.demo.models.Product;
+import com.example.demo.models.ProductList;
+import com.example.demo.models.ProductListItem;
+import com.example.demo.models.ProductListItemId;
 
 import org.junit.jupiter.api.Test;
 public class ProductListItemTest {
