+++ api/src/test/java/com/example/demo/model/SavedListTest.java
@@ -1,15 +1,8 @@
 package com.example.demo.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.SavedList;
+import com.example.demo.models.ProductList;
+import com.example.demo.models.SavedList;
 
 import org.junit.jupiter.api.Test;
 public class SavedListTest {
@@ -21,7 +14,6 @@ public class SavedListTest {
         list.setProductList(new ProductList());
         assertEquals(1L, list.getId());
         assertEquals("Pedrofilo", list.getName());
-        assertEquals(null, list.getProductList().getOrderList());
     }
 
     @Test
@@ -30,6 +22,5 @@ public class SavedListTest {
         list.setId(1L);
         assertEquals(1L, list.getId());
         assertEquals("Pedrofilo", list.getName());
-        assertEquals(null, list.getProductList().getOrderList());
     }
 }
