+++ api/src/test/java/com/example/demo/model/StoreTest.java
@@ -1,18 +1,8 @@
 package com.example.demo.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import com.example.demo.Models.Address;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.SavedList;
-import com.example.demo.Models.Store;
+import com.example.demo.models.Address;
+import com.example.demo.models.Store;
 
 import org.junit.jupiter.api.Test;
 public class StoreTest {
@@ -28,20 +18,17 @@ public class StoreTest {
         store.setName("Continente");
         assertEquals(1L, store.getId());
         assertEquals("Continente", store.getName());
-            assertEquals(0, store.getOrderList().size());
     }
     
 
 	@Test
     void testWhenCreateValidStoreConstructorThenReturnStore(){
-        Set<OrderList> orderList = new HashSet<>();
         Address address=  new Address();
         address.setCity("Aveiro");
-        Store store= new Store("Continente",address, orderList);
+        Store store= new Store("Continente", address);
         store.setId(1L);
         assertEquals(1L, store.getId());
         assertEquals("Continente", store.getName());
-            assertEquals(0, store.getOrderList().size());
     }
     
 }
