+++ api/src/test/java/com/example/demo/repository/CategoryRepositoryTest.java
@@ -1,44 +1,26 @@
 package com.example.demo.repository;
-import javax.persistence.PersistenceException;
+
 import javax.validation.ConstraintViolationException;
 
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.User;
-import com.example.demo.Models.UserAddress;
-import com.example.demo.Repository.AddressRepository;
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Repository.CategoryRepository;
-import com.example.demo.Repository.ProductListRepository;
-import com.example.demo.Repository.UserRepository;
+import com.example.demo.models.Category;
+import com.example.demo.repository.CategoryRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 
-import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @DataJpaTest
 @Testcontainers
@@ -65,9 +47,7 @@ public class CategoryRepositoryTest {
 
     @Test
     void testWhenCreateCategoryAndFindById_thenReturnSameCategory() {
-        Set<Product> x = new HashSet();
-        Category cat = new Category("Vegetais", false, x);
-        
+        Category cat = new Category("Vegetais", false);
         
         entityManager.persistAndFlush(cat);
         Optional<Category> res = rep.findById(cat.getId());
@@ -86,11 +66,8 @@ public class CategoryRepositoryTest {
 
     @Test
     void testGivenAddressAndFindByAll_thenReturnSameAddress() {
-        Set<Product> x = new HashSet();
-        Set<Product> y = new HashSet();
-
-        Category cat = new Category("Vegetais", false, x);
-        Category cat2 = new Category("Fruta", false, y);
+        Category cat = new Category("Vegetais", false);
+        Category cat2 = new Category("Fruta", false);
         entityManager.persistAndFlush(cat);
         entityManager.persistAndFlush(cat2);
         List<Category> all = rep.findAll();
