+++ api/src/test/java/com/example/demo/repository/ProductListItemRepositoryTest.java
@@ -1,53 +1,36 @@
 package com.example.demo.repository;
 
-import com.example.demo.Repository.ProductListItemRepository;
-import javax.persistence.PersistenceException;
-import javax.validation.ConstraintViolationException;
+import com.example.demo.models.Address;
+import com.example.demo.models.Category;
+import com.example.demo.models.OrderList;
+import com.example.demo.models.Product;
+import com.example.demo.models.ProductList;
+import com.example.demo.models.ProductListItem;
+import com.example.demo.models.ProductListItemId;
+import com.example.demo.models.Store;
+import com.example.demo.models.User;
+import com.example.demo.models.UserAddress;
+import com.example.demo.repository.ProductListItemRepository;
 
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.OrderProductItemId;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.ProductListItemId;
-import com.example.demo.Models.Store;
-import com.example.demo.Models.User;
-import com.example.demo.Models.UserAddress;
-import com.example.demo.Repository.AddressRepository;
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Repository.CategoryRepository;
-import com.example.demo.Repository.OrderProductItemRepository;
-import com.example.demo.Repository.ProductListRepository;
-import com.example.demo.Repository.UserRepository;
+import javax.persistence.PersistenceException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 
 import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @DataJpaTest
 @Testcontainers
@@ -77,21 +60,16 @@ public class ProductListItemRepositoryTest {
     void testWhenCreateOrderProductItemAndFindById_thenReturnSameOrderProductItem() {
         Store store = new Store();
         store.setName("puma");
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
-        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas", store);
+        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas");
         UserAddress userAddressTmp = new UserAddress(user, address);
-        Set<OrderProductItem> orderProductItem=new HashSet<>();
         ProductList productListTmp = new ProductList();
         productListTmp.setUser(user);
-        Set<OrderProductItem> productListItems=new HashSet<>();
         Long deliveryTimestamp=111111111111L;
         store.setAddress(address);
         Long deliveryId=1L;
-        OrderList list = new OrderList(productListTmp, orderProductItem, address, store, deliveryId, deliveryTimestamp);
-        list.setProductListItems(productListItems);
+        OrderList list = new OrderList(productListTmp, address, store, deliveryId, deliveryTimestamp);
         entityManager.persistAndFlush(user);
         entityManager.persistAndFlush(store);
         entityManager.persistAndFlush(address);
@@ -99,11 +77,9 @@ public class ProductListItemRepositoryTest {
         entityManager.persistAndFlush(productListTmp);
         entityManager.persistAndFlush(list);
 
-        Set<Product> x = new HashSet();
-        Category cat = new Category("Vegetais", false, x);
+        Category cat = new Category("Vegetais", false);
         
         Product product = new Product("Pilhas", 5.1f, "leve", true, cat);
-        x.add(product);
         
         entityManager.persistAndFlush(cat);
         entityManager.persistAndFlush(product);
@@ -136,21 +112,16 @@ public class ProductListItemRepositoryTest {
     void testGivenProductListItemAndFindByAll_thenReturnSameProductListItem() {
         Store store = new Store();
         store.setName("puma");
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
-        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas", store);
+        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas");
         UserAddress userAddressTmp = new UserAddress(user, address);
         ProductList productListTmp = new ProductList();
-        Set<OrderProductItem> orderProductItem=new HashSet<>();
         productListTmp.setUser(user);
-        Set<OrderProductItem> productListItems=new HashSet<>();
         Long deliveryTimestamp=111111111111L;
         store.setAddress(address);
         Long deliveryId=1L;
-        OrderList list = new OrderList(productListTmp, orderProductItem, address, store, deliveryId, deliveryTimestamp);
-        list.setProductListItems(productListItems);
+        OrderList list = new OrderList(productListTmp, address, store, deliveryId, deliveryTimestamp);
         entityManager.persistAndFlush(user);
         entityManager.persistAndFlush(store);
         entityManager.persistAndFlush(address);
@@ -158,15 +129,10 @@ public class ProductListItemRepositoryTest {
         entityManager.persistAndFlush(productListTmp);
         entityManager.persistAndFlush(list);
 
-        Set<Product> x = new HashSet();
-        Category cat = new Category("Vegetais", false, x);
+        Category cat = new Category("Vegetais", false);
 
         Product product = new Product("Pilhas", 5.1f, "leve", true, cat);
         Product product2 = new Product("Pilhas reciclaveis", 12, "leve", true, cat);
-       
-       
-        x.add(product);
-        x.add(product2);
         
         entityManager.persistAndFlush(cat);
         entityManager.persistAndFlush(product);
