+++ api/src/test/java/com/example/demo/repository/ProductListRepositoryTest.java
@@ -1,52 +1,28 @@
 package com.example.demo.repository;
-import com.example.demo.Repository.ProductListItemRepository;
-import javax.persistence.PersistenceException;
-import javax.validation.ConstraintViolationException;
 
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.OrderProductItemId;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.ProductListItemId;
-import com.example.demo.Models.Store;
-import com.example.demo.Models.User;
-import com.example.demo.Models.UserAddress;
-import com.example.demo.Repository.AddressRepository;
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Repository.CategoryRepository;
-import com.example.demo.Repository.OrderProductItemRepository;
-import com.example.demo.Repository.ProductListRepository;
-import com.example.demo.Repository.UserRepository;
+import javax.persistence.PersistenceException;
+
+import com.example.demo.models.ProductList;
+import com.example.demo.models.User;
+import com.example.demo.repository.ProductListRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 
 import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @DataJpaTest
 @Testcontainers
@@ -73,9 +49,7 @@ public class ProductListRepositoryTest {
 
     @Test
     void testWhenCreateProductListAndFindById_thenReturnSameProductList() {
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
         ProductList productListTmp = new ProductList();
         productListTmp.setUser(user);
@@ -99,17 +73,13 @@ public class ProductListRepositoryTest {
 
     @Test
     void testGivenProductListItemAndFindByAll_thenReturnSameProductListItem() {
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
         ProductList productListTmp = new ProductList();
         productListTmp.setUser(user);
         entityManager.persistAndFlush(productListTmp);
 
-        Set<ProductList> productList1 = new HashSet<>();
-        Set<UserAddress> userAddress1=new HashSet<>();
-        User user1 = new User("ada@gmail.com", "Ser1111ras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList1, userAddress1);
+        User user1 = new User("ada@gmail.com", "Ser1111ras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user1);
         ProductList productListTmp1 = new ProductList();
         productListTmp1.setUser(user1);
