+++ api/src/test/java/com/example/demo/repository/SavedListRepositoryTest.java
@@ -1,43 +1,29 @@
 package com.example.demo.repository;
-import javax.persistence.PersistenceException;
-import javax.validation.ConstraintViolationException;
 
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.SavedList;
-import com.example.demo.Models.User;
-import com.example.demo.Models.UserAddress;
-import com.example.demo.Repository.AddressRepository;
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Repository.ProductListRepository;
-import com.example.demo.Repository.SavedListRepository;
-import com.example.demo.Repository.UserRepository;
+import javax.persistence.PersistenceException;
+
+import com.example.demo.models.ProductList;
+import com.example.demo.models.SavedList;
+import com.example.demo.models.User;
+import com.example.demo.repository.SavedListRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 
 import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @DataJpaTest
 @Testcontainers
@@ -68,9 +54,7 @@ public class SavedListRepositoryTest {
     void testWhenCreateSavedListAndFindById_thenReturnSameSavedList() {
         SavedList x = new SavedList();
         ProductList list = new ProductList();
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
         list.setUser(user);
         entityManager.persistAndFlush(list);
@@ -100,9 +84,7 @@ public class SavedListRepositoryTest {
         ProductList list = new ProductList();
         ProductList list2 = new ProductList();
 
-        Set<ProductList> productList = new HashSet<>();
-        Set<UserAddress> userAddress=new HashSet<>();
-        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true, productList, userAddress);
+        User user = new User("alex200020011@gmail.com", "Serras", "aaaaa", new Date(2000, 5, 28), "911912912", false, true);
         entityManager.persistAndFlush(user);
         list.setUser(user);
         list2.setUser(user);
