+++ api/src/test/java/com/example/demo/repository/StoreRepositoryTest.java
@@ -1,54 +1,30 @@
 package com.example.demo.repository;
-import com.example.demo.Repository.ProductListItemRepository;
-import javax.persistence.PersistenceException;
+
 import javax.validation.ConstraintViolationException;
 
-import com.example.demo.Models.Address;
-import com.example.demo.Models.CartList;
-import com.example.demo.Models.Category;
-import com.example.demo.Models.OrderList;
-import com.example.demo.Models.OrderProductItem;
-import com.example.demo.Models.OrderProductItemId;
-import com.example.demo.Models.Product;
-import com.example.demo.Models.ProductList;
-import com.example.demo.Models.ProductListItem;
-import com.example.demo.Models.ProductListItemId;
-import com.example.demo.Models.Store;
-import com.example.demo.Models.User;
-import com.example.demo.Models.UserAddress;
-import com.example.demo.Repository.AddressRepository;
-import com.example.demo.Repository.CartListRepository;
-import com.example.demo.Repository.CategoryRepository;
-import com.example.demo.Repository.OrderProductItemRepository;
-import com.example.demo.Repository.ProductListRepository;
-import com.example.demo.Repository.ProductRepository;
-import com.example.demo.Repository.StoreRepository;
-import com.example.demo.Repository.UserRepository;
+import com.example.demo.models.Address;
+import com.example.demo.models.Store;
+import com.example.demo.repository.StoreRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertThrows;
 
-import java.util.Date;
-import java.util.HashSet;
+
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
-import org.junit.jupiter.api.MethodOrderer;
+
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.TestPropertySource;
 import org.testcontainers.containers.MySQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 @DataJpaTest
 @Testcontainers
@@ -78,7 +54,7 @@ public class StoreRepositoryTest {
 
         Store store = new Store();
         store.setName("puma");
-        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas", store);
+        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas");
         store.setAddress(address);
         entityManager.persistAndFlush(store);
 
@@ -97,18 +73,16 @@ public class StoreRepositoryTest {
 
     @Test
     void testGivenAddressAndFindByAll_thenReturnSameAddress() {
-        
-
         Store store = new Store();
         store.setName("puma");
-        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas", store);
+        Address address= new Address("Portugal", "1903-221", "Aveiro", "Rua das Pombas");
         store.setAddress(address);
         entityManager.persistAndFlush(store);
 
 
         Store store1 = new Store();
         store1.setName("JOMA");
-        Address address1= new Address("Portugal", "3903-221", "Santarem", "Rua das Estia", store1);
+        Address address1= new Address("Portugal", "3903-221", "Santarem", "Rua das Estia");
         store1.setAddress(address1);
         entityManager.persistAndFlush(store1);
